+++ src/app/api/technologies/route.ts
@@ -1,9 +1,25 @@
 import { NextResponse, NextRequest } from "next/server";
 import checkkeys from "@/lib/checkKeys";
+import { headers } from "next/headers";
+import { decode } from "@/lib/jwt";
 import prisma from "@/lib/prisma";
 
 export async function POST(request: Request) {
   const json = await request.json();
+  const Authentication = headers().get("Authentication") || "";
+  const decoded = decode(Authentication);
+
+  if (!decoded.success) {
+    return NextResponse.json(
+      {
+        status: 401,
+        message: "The login information is incorrect. Please log in again.",
+      },
+      {
+        status: 401,
+      }
+    );
+  }
 
   const requiredKeys = ["name", "image"];
 
@@ -68,6 +84,20 @@ export async function GET(request: NextRequest) {
 
 export async function PUT(request: Request) {
   const json = await request.json();
+  const Authentication = headers().get("Authentication") || "";
+  const decoded = decode(Authentication);
+
+  if (!decoded.success) {
+    return NextResponse.json(
+      {
+        status: 401,
+        message: "The login information is incorrect. Please log in again.",
+      },
+      {
+        status: 401,
+      }
+    );
+  }
 
   const requiredKeys = [
     "id",
@@ -106,6 +136,20 @@ export async function PUT(request: Request) {
 
 export async function DELETE(request: NextRequest) {
   const id = request.nextUrl.searchParams.get("id") || "";
+  const Authentication = headers().get("Authentication") || "";
+  const decoded = decode(Authentication);
+
+  if (!decoded.success) {
+    return NextResponse.json(
+      {
+        status: 401,
+        message: "The login information is incorrect. Please log in again.",
+      },
+      {
+        status: 401,
+      }
+    );
+  }
 
   try {
     const technology = await prisma.technology.delete({
